step one move to cra
