+++ package.json
@@ -1,34 +1,25 @@
 {
-  "name": "performance-poc",
+  "name": "poc-gcp-compression-frontend",
   "version": "1.0.0",
-  "description": "Highly performant web application",
+  "description": "PoC of GCP Compression frontend",
   "main": "index.js",
   "scripts": {
-    "test": "eslint src && preact test",
-    "start": "if-env NODE_ENV=production && npm run -s serve || npm run -s dev",
-    "build": "preact build",
-    "serve": "preact build && preact serve",
-    "dev": "preact watch"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "eslintConfig": {
-    "extends": "eslint-config-synacor"
-  },
-  "devDependencies": {
-    "eslint": "^4.5.0",
-    "eslint-config-synacor": "^1.1.0",
-    "if-env": "^1.0.0",
-    "preact-cli": "^1.4.1"
-  },
   "dependencies": {
     "instantly": "^1.4.1",
     "isomorphic-fetch": "^2.2.1",
     "lodash.chunk": "^4.2.0",
-    "preact": "^8.2.1",
-    "preact-compat": "^3.17.0",
-    "preact-router": "^2.5.7",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-router-dom": "^4.2.2",
+    "react-scripts": "^1.0.14",
     "semantic-ui-css": "^2.2.12",
     "semantic-ui-react": "^0.73.0",
     "unfetch": "^3.0.0"
