+++ src/index.js
@@ -1,6 +1,6 @@
-import { h, render } from 'preact';
-
-import './style';
+import React from 'react';
+import ReactDOM from 'react-dom';
 import App from './components/app';
+import './style/index.css';
 
-export default () => (<App />);
+ReactDOM.render(<App />, document.getElementById('root'));
